Add datasets download to Dockerfile

+++ setup_datasets.sh
@@ -28,8 +28,14 @@ function download_file_zst() {
          wget -O - ${URL} | zstd -d > ${FILE}
       fi
    else
-      # Does not exists -> download
-      wget -O - ${URL} | zstd -d > ${FILE}
+      # Check if downsampled version exists
+      DOWNSAMPLED_FILE=${FILE/800M/200M}
+      if [ ! -f DOWNSAMPLED_FILE ]; then
+         # Does not exists -> download
+         wget -O - ${URL} | zstd -d > ${FILE}
+      else
+         return
+      fi
    fi
 
    # Validate (at this point the file should really exist)
@@ -47,7 +53,7 @@ function download_file_zst() {
 
 # Check if the user has provided an argument
 if [ $# -eq 0 ]; then
-    echo "Using default data directory: ./data"
+    echo "Using default directory: ./data"
     new_directory="./data"
 else
     new_directory=$1
@@ -70,8 +76,8 @@ cd $new_directory
 # Format: download_file <file_name> <md5_checksum> <url>
 download_file_zst wiki_ts_200M_uint64 4f1402b1c476d67f77d2da4955432f7d https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/JGVF9A/SVN8PI
 # download_file_zst books_200M_uint32 9f3e578671e5c0348cdddc9c68946770 https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/JGVF9A/5YTV8K  
-download_file_zst osm_cellids_800M_uint64 70670bf41196b9591e07d0128a281b9a https://www.dropbox.com/s/j1d4ufn4fyb4po2/osm_cellids_800M_uint64.zst?dl=1
-download_file_zst fb_200M_uint64 3b0f820caa0d62150e87ce94ec989978 https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/JGVF9A/EATHF7  
+download_file_zst fb_200M_uint64 3b0f820caa0d62150e87ce94ec989978 https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/JGVF9A/EATHF7 
+download_file_zst osm_cellids_800M_uint64 70670bf41196b9591e07d0128a281b9a https://www.dropbox.com/s/j1d4ufn4fyb4po2/osm_cellids_800M_uint64.zst?dl=1 
 
 cd $initial_dir
 echo "done"
